+++ CHANGES.md
@@ -18,15 +18,15 @@
 * [TODO] Add support for variable/multi positional params `{:? [vals...]}` in named params
 * [TODO] Do not swallow exception that caused rollback attempt in a transaction.
 * Params setter utility in `asphalt.param` namespace
-  * Macros `lay-params`, `lay-params-vec`, `lay-params-map` with support for
+  * Macro `lay-params` with support for
     * Available at macro-expansion time (efficient)
     * Param types (single-value and multi-value)
-    * Param position/name reference
+    * Param position/name reference (vector/map/nil params)
     * Additional coercion arguments
   * Function `set-params` with run time support for
     * Available at run time (flexible)
     * Param types (single-value and multi-value)
-    * Param position/name reference
+    * Param position/name reference (vector/map/nil params)
     * Additional coercion arguments
 * Queries
   * [BREAKING CHANGE] Accept options in `asphalt.core` fetch fns (arity 3)